Addressed review comments by moving interfaces and enums to separate packages.
