+++ document-store/src/main/java/org/hypertrace/core/documentstore/expression/ConstantExpression.java
@@ -4,6 +4,7 @@ import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Value;
+import org.hypertrace.core.documentstore.expression.type.SelectingExpression;
 import org.hypertrace.core.documentstore.parser.SelectingExpressionParser;
 
 /**
