+++ document-store/src/main/java/org/hypertrace/core/documentstore/expression/IdentifierExpression.java
@@ -1,6 +1,9 @@
 package org.hypertrace.core.documentstore.expression;
 
 import lombok.Value;
+import org.hypertrace.core.documentstore.expression.type.GroupingExpression;
+import org.hypertrace.core.documentstore.expression.type.SelectingExpression;
+import org.hypertrace.core.documentstore.expression.type.SortingExpression;
 import org.hypertrace.core.documentstore.parser.GroupingExpressionParser;
 import org.hypertrace.core.documentstore.parser.SelectingExpressionParser;
 import org.hypertrace.core.documentstore.parser.SortingExpressionParser;
