--- document-store/src/main/java/org/hypertrace/core/documentstore/expression/LogicalOperator.java
@@ -1,6 +1,0 @@
-package org.hypertrace.core.documentstore.expression;
-
-public enum LogicalOperator {
-  AND,
-  OR,
-}
