+++ document-store/src/main/java/org/hypertrace/core/documentstore/expression/RelationalExpression.java
@@ -1,6 +1,9 @@
 package org.hypertrace.core.documentstore.expression;
 
 import lombok.Value;
+import org.hypertrace.core.documentstore.expression.operators.RelationalOperator;
+import org.hypertrace.core.documentstore.expression.type.FilteringExpression;
+import org.hypertrace.core.documentstore.expression.type.SelectingExpression;
 import org.hypertrace.core.documentstore.parser.FilteringExpressionParser;
 
 /**
