--- document-store/src/main/java/org/hypertrace/core/documentstore/expression/SortingOrder.java
@@ -1,6 +1,0 @@
-package org.hypertrace.core.documentstore.expression;
-
-public enum SortingOrder {
-  ASC,
-  DESC,
-}
