+++ document-store/src/main/java/org/hypertrace/core/documentstore/expression/operators/LogicalOperator.java
@@ -1,0 +1,6 @@
+package org.hypertrace.core.documentstore.expression.operators;
+
+public enum LogicalOperator {
+  AND,
+  OR,
+}
