+++ document-store/src/main/java/org/hypertrace/core/documentstore/expression/operators/RelationalOperator.java
@@ -1,4 +1,4 @@
-package org.hypertrace.core.documentstore.expression;
+package org.hypertrace.core.documentstore.expression.operators;
 
 public enum RelationalOperator {
   EQ,
