+++ document-store/src/main/java/org/hypertrace/core/documentstore/expression/operators/SortingOrder.java
@@ -1,0 +1,6 @@
+package org.hypertrace.core.documentstore.expression.operators;
+
+public enum SortingOrder {
+  ASC,
+  DESC,
+}
