+++ document-store/src/main/java/org/hypertrace/core/documentstore/query/SortingDefinition.java
@@ -1,8 +1,8 @@
 package org.hypertrace.core.documentstore.query;
 
 import lombok.Value;
-import org.hypertrace.core.documentstore.expression.SortingExpression;
-import org.hypertrace.core.documentstore.expression.SortingOrder;
+import org.hypertrace.core.documentstore.expression.operators.SortingOrder;
+import org.hypertrace.core.documentstore.expression.type.SortingExpression;
 
 /**
  * A generic ORDER BY definition that supports expressions. Note that this class is a more general
